backend/requirements.txt
django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
django-filter==25.1
Pillow==10.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0